# api_gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Messenger API",
        "contact": {},
        "version": "0.2"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Авторизация",
                "parameters": [
                    {
                        "description": "Данные для входа",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Login"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/my_profile": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление пароля",
                "responses": {}
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Refresh",
                "parameters": [
                    {
                        "description": "Обновление токена",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Refresh"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Register"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/reset_password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Подтвердить OTP код",
                "parameters": [
                    {
                        "description": "Данные для проверки кода",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPassword"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/send_otp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Отправить код на почту",
                "parameters": [
                    {
                        "description": "Данные для отправки кода на почту",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SendOtp"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/auth/update_password": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление пароля",
                "parameters": [
                    {
                        "description": "Данные для обновления пароля",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdatePassword"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/chat/get_chats": {
            "get": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chat"
                ],
                "summary": "получение списка чатов",
                "responses": {}
            }
        },
        "/ping": {
            "get": {
                "description": "Публичный запрос, который клиент отправляет при запуске. Возвращает статус сервисов.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ping"
                ],
                "summary": "Проверка доступности сервера и версии приложения",
                "responses": {}
            }
        },
        "/profile/info": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Получение профиля пользователя по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Username пользователя",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/profile/update_bio": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить bio",
                "parameters": [
                    {
                        "description": "Данные для обновления профиля",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProfile"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/profile/update_email": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить nickname",
                "parameters": [
                    {
                        "description": "Данные для обновления профиля",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProfile"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/profile/update_nickname": {
            "post": {
                "security": [
                    {
                        "Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "profile"
                ],
                "summary": "Обновить nickname",
                "parameters": [
                    {
                        "description": "Данные для обновления профиля",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateProfile"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/search/profiles": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "search"
                ],
                "summary": "Поиск пользователей по username",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Username пользователя",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/ws": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "web_socket"
                ],
                "summary": "WebSocket connect",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT access token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "Switching Protocols",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Login": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Refresh": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                }
            }
        },
        "models.Register": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.ResetPassword": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "models.SendOtp": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.UpdatePassword": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "reset_token": {
                    "type": "string"
                }
            }
        },
        "models.UpdateProfile": {
            "type": "object",
            "properties": {
                "value": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Token": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}